Added 2 files.
isingpackage.py: contains the class spin_array to hold the information of the Ising model (J, kb, H etc).
The class includes pretty much all the methods we need to control the model.

ising_model_project.ipynb: contains the scripts that uses the spin_array to investigate the 2D Ising model.
Has succesfully ran the 64*64 model for 100 steps, and generated an animation of the spin array.
Also has the (Energy vs step) plot to investigate the time spent to reach equilibrium.